// api/models/CourseChapterView.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CourseChapterContentView } from './CourseChapterContentView';
export type CourseChapterView = {
    id: number;
    name: string;
    index: number;
    course_id: number;
    chapter_contents: Array<CourseChapterContentView>;
    created_at: string;
    updated_at: string;
};